+++ library/src/utils/view-transtions.ts
@@ -1,3 +1,5 @@
+import type { Modifiers } from "../engine/types"
+
 export interface DocumentSupportingViewTransitionAPI {
   startViewTransition(
     updateCallback: () => Promise<void> | void,
@@ -13,5 +15,19 @@ export interface IViewTransition {
 
 export const docWithViewTransitionAPI =
   document as unknown as DocumentSupportingViewTransitionAPI
+
 export const supportsViewTransitions =
   !!docWithViewTransitionAPI.startViewTransition
+
+export function modifyViewTransition(
+  callback: (...args: any[]) => any,
+  mods: Modifiers,
+  ) {
+  if (mods.has('viewtransition') && supportsViewTransitions) {
+    const cb = callback // I hate javascript
+    callback = (...args: any[]) =>
+      document.startViewTransition(() => cb(...args))
+  }
+
+  return callback
+}
